Clean up build, fix some compile errors

+++ src/memoryusage.h
@@ -2,15 +2,16 @@
 #define LTTLWHWHT_MEMORYUSAGE_H
 
 #include <stdlib.h>
+#include <stdint.h>
 
 class MemoryUsage {
   public:
     virtual ~MemoryUsage() {}
     virtual bool tryUpdate() = 0;
-    virtual size_t getVSS() const = 0;
-    virtual size_t getRSS() const = 0;
-    virtual size_t getPSS() const = 0;
-    virtual size_t getUSS() const = 0;
+    virtual uint64_t getVSS() const = 0;
+    virtual uint64_t getRSS() const = 0;
+    virtual uint64_t getPSS() const = 0;
+    virtual uint64_t getUSS() const = 0;
 };
 
 #endif // LTTLWHWHT_MEMORYUSAGE_H
